requirements.txt
# ============================================
# AI-Powered AML System - Enterprise Edition
# Complete Requirements with KYC, Sanctions Screening, Law Enforcement Integration
# ============================================

# Core Django Framework
Django==5.1.7
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
channels==4.1.0  # WebSocket support for real-time updates
daphne==4.1.2  # ASGI server

# Database & Caching
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0
celery==5.4.0  # Async task processing

# Authentication & Security
djangorestframework-simplejwt==5.4.0
django-otp==1.5.4  # Multi-factor authentication
qrcode==8.0  # For MFA QR codes
cryptography==44.0.0  # Encryption utilities
pyjwt==2.10.1

# Machine Learning & AI
numpy==2.2.1
scikit-learn==1.6.1
pandas==2.2.3
xgboost==2.1.3  # Gradient boosting
lightgbm==4.5.0  # Light gradient boosting
imbalanced-learn==0.12.4  # Handle imbalanced datasets
shap==0.46.0  # Explainable AI (SHAP values)
lime==0.2.0.1  # Local interpretable model explanations

# Graph Analysis & Network Detection
networkx==3.4.2  # Graph algorithms
python-louvain==0.16  # Community detection
igraph==0.11.8  # Fast graph library
torch==2.5.1  # For graph neural networks
torch-geometric==2.6.1  # Graph neural network library

# Fuzzy Matching & Text Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1  # Fast fuzzy string matching
jellyfish==1.1.0  # Phonetic matching algorithms
rapidfuzz==3.10.1  # Fast fuzzy matching

# Document Processing & OCR
Pillow==11.0.0  # Image processing
pdf2image==1.17.0  # PDF to image conversion
pytesseract==0.3.13  # OCR for document verification
python-magic==0.4.27  # File type detection

# Data Validation & Serialization
marshmallow==3.23.2
pydantic==2.10.4  # Data validation
phonenumbers==8.13.50  # Phone number validation

# API Integrations
requests==2.32.3
urllib3==2.2.3
httpx==0.28.1  # Async HTTP client

# File Storage & Handling
boto3==1.35.84  # AWS S3 integration
django-storages==1.14.4  # Cloud storage backends
minio==7.2.11  # MinIO object storage

# Reporting & Export
reportlab==4.2.5  # PDF generation
openpyxl==3.1.5  # Excel export
python-docx==1.1.2  # Word document generation
WeasyPrint==63.1  # HTML to PDF

# Forensics & Digital Signatures
hashlib-additional==1.0  # Additional hash algorithms
ecdsa==0.19.0  # Digital signatures
python-gnupg==0.5.3  # GPG encryption

# Monitoring & Logging
sentry-sdk==2.19.2  # Error tracking
django-prometheus==2.3.1  # Metrics export
python-json-logger==3.2.1  # JSON logging
elastic-apm==6.23.0  # Application performance monitoring

# Rate Limiting & Throttling
django-ratelimit==4.1.0

# Task Scheduling
celery-beat==2.7.0  # Periodic tasks
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Data Migration & ETL
great-expectations==1.2.6  # Data validation framework

# Geolocation
geoip2==4.8.1  # IP geolocation
geopy==2.4.1  # Geocoding

# Utilities
python-dateutil==2.9.0
pytz==2025.1
arrow==1.3.0  # Better datetime handling
python-dotenv==1.0.1  # Environment variables
pyyaml==6.0.2  # YAML parsing

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0
factory-boy==3.3.1  # Test data factories
faker==33.1.0  # Fake data generation
locust==2.32.5  # Load testing

# Code Quality
black==24.10.0
flake8==7.1.1
pylint==3.3.2
mypy==1.14.0  # Type checking
bandit==1.8.0  # Security linting

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2
uvicorn==0.34.0  # ASGI server

# Backup & Database Tools
django-dbbackup==4.2.1

# WebSocket & Real-time
channels-redis==4.2.1

# API Documentation
drf-spectacular==0.28.0  # OpenAPI/Swagger generation

# Compliance & Audit
django-simple-history==3.8.0  # Model history tracking
django-auditlog==3.1.0  # Audit logging

